# Draft: Faster `wrapdims` over tables