@rofinn 2022-08-31
Closing in favour of #126 which reduces the allocations by another order of magnitude and is faster in all cases I tried.